allow shell pipelines
in fx:command strings
